+++ src/main/java/org/cloudwarp/probablychests/world/feature/NormalChestFeature.java
@@ -26,7 +26,7 @@ public class NormalChestFeature extends Feature<PCChestFeatureConfig> {
 		PCChestFeatureConfig config = context.getConfig();
 		structureWorldAccess.setBlockState(pos, PCBlocks.NORMAL_CHEST.getDefaultState(), 3);
 		//structureWorldAccess.setBlockState(pos, Blocks.GLOWSTONE.getDefaultState(), 3);
-		LootableContainerBlockEntity.setLootTable(structureWorldAccess, random, pos, PCLootTables.LUSH_CHEST);
+		LootableContainerBlockEntity.setLootTable(structureWorldAccess, random, pos, PCLootTables.NORMAL_CHEST);
 
 		return true;
 	}