+++ src/main/java/org/cloudwarp/probablychests/world/feature/PCGroundPlacementModifier.java
@@ -21,7 +21,6 @@ public class PCGroundPlacementModifier extends PlacementModifier {
 	public static final Codec<PCGroundPlacementModifier> MODIFIER_CODEC = RecordCodecBuilder.create((instance) -> instance.group(
 					(Direction.VERTICAL_CODEC.fieldOf("direction_of_search")).forGetter(PCGroundPlacementModifier -> PCGroundPlacementModifier.direction),
 					(BlockPredicate.BASE_CODEC.fieldOf("direction_target_condition")).forGetter(PCGroundPlacementModifier -> PCGroundPlacementModifier.directionPredicate),
-					(BlockPredicate.BASE_CODEC.fieldOf("direction_target_block_condition")).forGetter(PCGroundPlacementModifier -> PCGroundPlacementModifier.directionBlockPredicate),
 					(BlockPredicate.BASE_CODEC.fieldOf("target_condition")).forGetter(PCGroundPlacementModifier -> PCGroundPlacementModifier.targetPredicate),
 					(Codec.intRange(1, 32).fieldOf("max_steps")).forGetter(PCGroundPlacementModifier -> PCGroundPlacementModifier.maxSteps),
 					Heightmap.Type.CODEC.fieldOf("heightmap").forGetter((PCGroundPlacementModifier) -> PCGroundPlacementModifier.heightmap))
@@ -29,36 +28,34 @@ public class PCGroundPlacementModifier extends PlacementModifier {
 	//---------------------------------------------------
 	private final Direction direction;
 	private final BlockPredicate directionPredicate;
-	private final BlockPredicate directionBlockPredicate;
 	private final BlockPredicate targetPredicate;
 	private final int maxSteps;
 	private final Heightmap.Type heightmap;
 
 
-	private PCGroundPlacementModifier (Direction direction, BlockPredicate directionPredicate, BlockPredicate directionBlockPredicate, BlockPredicate targetPredicate, int maxSteps, Heightmap.Type heightmap) {
+	private PCGroundPlacementModifier (Direction direction, BlockPredicate directionPredicate, BlockPredicate targetPredicate, int maxSteps, Heightmap.Type heightmap) {
 		this.direction = direction;
 		this.targetPredicate = targetPredicate;
 		this.directionPredicate = directionPredicate;
-		this.directionBlockPredicate = directionBlockPredicate;
 		this.maxSteps = maxSteps;
 		this.heightmap = heightmap;
 	}
 
-	public static PCGroundPlacementModifier of (Direction direction, BlockPredicate directionPredicate, BlockPredicate directionBlockPredicate, BlockPredicate targetPredicate, int maxSteps, Heightmap.Type heightmap) {
-		return new PCGroundPlacementModifier(direction, directionPredicate, directionBlockPredicate, targetPredicate, maxSteps, heightmap);
+	public static PCGroundPlacementModifier of (Direction direction, BlockPredicate directionPredicate, BlockPredicate targetPredicate, int maxSteps, Heightmap.Type heightmap) {
+		return new PCGroundPlacementModifier(direction, directionPredicate, targetPredicate, maxSteps, heightmap);
 	}
 
 	@Override
 	public Stream<BlockPos> getPositions (FeaturePlacementContext context, Random random, BlockPos pos) {
 		BlockPos.Mutable mutableTarget = pos.mutableCopy();
-		int k = context.getTopY(this.heightmap, mutableTarget.getX(), mutableTarget.getZ());
-		mutableTarget.set(mutableTarget.getX(),k - (1 + random.nextInt(Math.abs(context.getBottomY() - k))),mutableTarget.getZ());
+		//int k = context.getTopY(this.heightmap, mutableTarget.getX(), mutableTarget.getZ());
+		//mutableTarget.set(mutableTarget.getX(),k - (1 + random.nextInt(Math.abs(context.getBottomY() - k))),mutableTarget.getZ());
 		BlockPos.Mutable mutableDirection = mutableTarget.mutableCopy();
 		mutableDirection.move(Direction.DOWN);
 		StructureWorldAccess structureWorldAccess = context.getWorld();
-		// &&
-		//					this.directionBlockPredicate.test(structureWorldAccess, mutableDirection)
 
+
+		// TODO: change IS AIR to REPLACABLE.
 		for (int i = 0; i < this.maxSteps; ++ i) {
 			if (this.targetPredicate.test(structureWorldAccess, mutableTarget) &&
 					this.directionPredicate.test(structureWorldAccess, mutableDirection) &&