+++ src/main/java/org/cloudwarp/probablychests/world/gen/LushChestGeneration.java
@@ -12,8 +12,12 @@ import java.util.List;
 
 public class LushChestGeneration {
 	public static void generateChest () {
-		BiomeModifications.addFeature(BiomeSelectors.foundInOverworld().and(BiomeSelectors.includeByKey(BiomeKeys.LUSH_CAVES)),
-		//BiomeModifications.addFeature(BiomeSelectors.all(),
+		BiomeModifications.addFeature(BiomeSelectors.all().and(BiomeSelectors.foundInOverworld()).and(context ->{
+			Biome biome = context.getBiome();
+			return biome.getTemperature() < 1.0f && biome.getTemperature() >= 0.5f;
+				}).and(BiomeSelectors.excludeByKey(List.of(BiomeKeys.DRIPSTONE_CAVES,BiomeKeys.OCEAN,BiomeKeys.COLD_OCEAN,BiomeKeys.DEEP_COLD_OCEAN,
+						BiomeKeys.DEEP_FROZEN_OCEAN,BiomeKeys.DEEP_OCEAN,BiomeKeys.DEEP_LUKEWARM_OCEAN,BiomeKeys.FROZEN_OCEAN,
+						BiomeKeys.LUKEWARM_OCEAN,BiomeKeys.WARM_OCEAN))),
 				GenerationStep.Feature.UNDERGROUND_STRUCTURES, PCFeatures.LUSH_CHEST_PLACED.getKey().get());
 	}
 }