+++ src/main/java/org/cloudwarp/probablychests/world/gen/NormalChestGeneration.java
@@ -1,7 +1,9 @@
 package org.cloudwarp.probablychests.world.gen;
 
 import net.fabricmc.fabric.api.biome.v1.BiomeModifications;
+import net.fabricmc.fabric.api.biome.v1.BiomeSelectionContext;
 import net.fabricmc.fabric.api.biome.v1.BiomeSelectors;
+import net.minecraft.world.biome.Biome;
 import net.minecraft.world.biome.BiomeKeys;
 import net.minecraft.world.gen.GenerationStep;
 import org.cloudwarp.probablychests.world.feature.PCFeatures;
@@ -10,8 +12,13 @@ import java.util.List;
 
 public class NormalChestGeneration {
 	public static void generateChest () {
-		BiomeModifications.addFeature(BiomeSelectors.foundInOverworld().and(BiomeSelectors.excludeByKey(List.of(BiomeKeys.LUSH_CAVES,BiomeKeys.DRIPSTONE_CAVES))),
-				//BiomeModifications.addFeature(BiomeSelectors.all(),
+		BiomeModifications.addFeature(BiomeSelectors.all().and(BiomeSelectors.foundInOverworld()).and(context -> {
+					Biome biome = context.getBiome();
+					return biome.getTemperature() <= 0.5f || (biome.getTemperature() == 0.8 && biome.getDownfall() == 0.4f);
+				}).and(BiomeSelectors.excludeByKey(List.of(BiomeKeys.LUSH_CAVES,BiomeKeys.DRIPSTONE_CAVES))),
+				GenerationStep.Feature.UNDERGROUND_STRUCTURES, PCFeatures.NORMAL_CHEST_PLACED.getKey().get());
+		//---------------------------------------------------
+		BiomeModifications.addFeature(BiomeSelectors.all().and(BiomeSelectors.foundInTheNether()),
 				GenerationStep.Feature.UNDERGROUND_STRUCTURES, PCFeatures.NORMAL_CHEST_PLACED.getKey().get());
 	}
 }