+++ src/main/java/org/cloudwarp/probablychests/world/gen/PCWorldGen.java
@@ -2,9 +2,9 @@ package org.cloudwarp.probablychests.world.gen;
 
 public class PCWorldGen {
 	public static void generatePCWorldGen () {
-		//LushChestGeneration.generateChest();
+		LushChestGeneration.generateChest();
 		NormalChestGeneration.generateChest();
-		//RockyChestGeneration.generateChest();
-		PCNormalPotGeneration.generatePot();
+		RockyChestGeneration.generateChest();
+		PCPotGeneration.generatePot();
 	}
 }