+++ changelog.md
@@ -6,3 +6,15 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Lush Chest
 - Clay Pot
+
+## [0.2.0] - 4/7/2022
+### Added
+- Normal Chest
+- Rocky Chest
+- Lush Pot
+- Rocky Pot
+- Nether Pot
+- New loot tables
+- Biome specific spawning
+- Spawn rate config
+- Made Pots silk touchable